fix!: tolerate v2 contracts in tss migration tests (#3002)
* fix!: tolerate v2 contracts in tss migration tests

* fix upgrade tests

---------

+++ pkg/contracts/testdappv2/TestDAppV2.sol
@@ -20,7 +20,7 @@ contract TestDAppV2 {
     struct RevertContext {
         address sender;
         address asset;
-        uint64 amount;
+        uint256 amount;
         bytes revertMessage;
     }
